[feat] HorizontalCollectionView와 HorizontalCollectionWithBackgroundImageView 분리
- 필수앱, 유료 순위, 무료 순위 섹션 구현
