--- 35-Seminar/Presentation/View/FinanceView/Cell/HorizontalTableViewCell.swift
@@ -1,12 +1,0 @@
-//
-//  HorizontalTableViewCell.swift
-//  35-Seminar
-//
-//  Created by OneTen on 10/30/24.
-//
-
-import UIKit
-
-class HorizontalTableViewCell: UITableViewCell {
-    
-}
