--- 35-Seminar/Presentation/View/FinanceView/HorizontalTableView.swift
@@ -1,12 +1,0 @@
-//
-//  HorizontalTableView.swift
-//  35-Seminar
-//
-//  Created by OneTen on 10/29/24.
-//
-
-import UIKit
-
-class HorizontalTableView: UITableViewCell {
-    
-}
